We put the method of obtaining roads into a separate class and interface
