+++ ServiceTelecom/Repositories/Interfaces/IRoadDataBase.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.ObjectModel;
+
+namespace ServiceTelecom.Repositories.Interfaces
+{
+    public interface IRoadDataBase
+    {
+        ObservableCollection<string> GetRoadDataBase(ObservableCollection<string> roadCollections);
+    }
+}
